--- Src/Pages/CurrentProvider/index.ts
@@ -1,1 +1,0 @@
-export { CurrentProvider } from "./CurrentProvider";
