+++ Src/Pages/Registration/ChooseAccountType.tsx
@@ -64,16 +64,16 @@ export const ChooseAccountType = ({
     setAppLoading(false);
 
     setTimeout(() => {
-      // navigation.dispatch(StackActions.replace('CurrentProvider'));
+      navigation.dispatch(StackActions.replace('CurrentProvider'));
 
       // if (identityResponse?.Identity?.status) {
       //   navigation.dispatch(StackActions.replace('AlmostThere'));
-      // } else if (userDetails?.email) {
+      // } else if (userDetails?.userName && userDetails?.email) {
       //   navigation.dispatch(StackActions.replace('TellUsAboutYou'));
       // } else if (userDetails?.userName && !userDetails?.email) {
       //   navigation.dispatch(StackActions.replace('EnterUserName'));
       // } else if (userDetails?.userName) {
-      //   navigation.dispatch(StackActions.replace('EnterEmail'));
+      //   navigation.dispatch(StackActions.replace('EnterUserName'));
       // } else if (defaultDomain) {
       //   navigation.dispatch(StackActions.replace('CurrentProvider'));
       // } else if (userDetails?.purpose) {
