+++ Src/Pages/Registration/CreatePin.tsx
@@ -20,11 +20,11 @@ import {
 import { ThemeContext } from '@Theme/Provider/ThemeContext';
 import { GlobalStyle as styles, TellUsAboutYouStyle } from '@Pages/Styles';
 import { NeuroTextInput } from '@Controls/NeuroTextInput';
-import { pinValidationSchema, Constants, computePinHash  } from '@Helpers/index';
-import { ThunkDispatch } from '@reduxjs/toolkit';
-import { useDispatch, useSelector } from 'react-redux';
+import { pinValidationSchema, Constants, computePinHash } from '@Helpers/index';
+import { useSelector } from 'react-redux';
 import { retrieveUserSession, storeUserSession } from '@Services/Storage';
-import {AgentAPI} from '../../Services/API/Agent';
+import { AgentAPI } from '../../Services/API/Agent';
+import { StackActions } from '@react-navigation/native';
 
 export const CreatePin = ({
   navigation,
@@ -34,8 +34,6 @@ export const CreatePin = ({
   const formikRef = useRef();
   const { userDetails } = useSelector((state) => state.user);
   const { identityResponse } = useSelector((state) => state.identity);
-  console.log(identityResponse.Identity.property)
-  const dispatch = useDispatch<ThunkDispatch<any, any, any>>();
 
   const newPin = useRef<TextInput>(null);
   const confirmPin = useRef<TextInput>(null);
@@ -43,7 +41,7 @@ export const CreatePin = ({
   if (userDetails) {
     Constants.UserDetails = userDetails;
     for (const item of identityResponse.Identity.property) {
-      console.log(item)
+      console.log(item);
       if (item.name === 'FIRST') {
         Constants.LegalIdentity.FIRST = item.value;
       }
@@ -67,18 +65,27 @@ export const CreatePin = ({
       objectId = AgentAPI.Account.getRandomValues(32);
       await storeUserSession(Constants.Authentication.ObjectId, objectId);
     }
-    const hashedPassword = computePinHash(password, objectId, Config.Host || '', userDetails.userName, userDetails.legalId);
+    const hashedPassword = computePinHash(
+      password,
+      objectId,
+      Config.Host || '',
+      userDetails.userName,
+      userDetails.legalId
+    );
     return hashedPassword;
   };
-  
+
   const handleFormSubmit = async (values: any) => {
     const hashedPassword = await hashPassword(values.newPin);
     await storeUserSession(Constants.Authentication.PinKey, hashedPassword);
-    const storedPassword = await retrieveUserSession(Constants.Authentication.PinKey);
+    const storedPassword = await retrieveUserSession(
+      Constants.Authentication.PinKey
+    );
     if (hashedPassword === storedPassword) {
-      alert(t('PIN.PinCreateSuccess'))
+      alert(t('PIN.PinCreateSuccess'));
+      navigation.dispatch(StackActions.replace('HomeScreen'));
     } else {
-      alert(t('PIN.IncorrectPin'))
+      alert(t('PIN.IncorrectPin'));
     }
   };
 
@@ -140,9 +147,7 @@ export const CreatePin = ({
           style={[TellUsAboutYouStyle(themeColors).detailHeight]}
           variant={TextLabelVariants.LABEL}
         >
-          {t(
-            'PIN.Description'
-          )}
+          {t('PIN.Description')}
         </TextLabel>
 
         <ScrollView showsVerticalScrollIndicator={false}>
@@ -251,7 +256,9 @@ export const CreatePin = ({
                   </TextLabel>
                 )}
 
-                <View style={TellUsAboutYouStyle(themeColors).actionButtonContainer}>
+                <View
+                  style={TellUsAboutYouStyle(themeColors).actionButtonContainer}
+                >
                   <ActionButton
                     disabled={!isValid}
                     textStyle={[TellUsAboutYouStyle(themeColors).sendText]}
