--- Src/Pages/TellUsAboutYou/index.ts
@@ -1,1 +1,0 @@
-export { TellUsAboutYou } from './TellUsAboutYou';
