--- Src/Pages/index.ts
@@ -1,2 +1,0 @@
-export { CurrentProvider } from './CurrentProvider';
-export { TellUsAboutYou } from './TellUsAboutYou';
